\encoding{latin1}

\name{ICSNP-package}
\alias{ICSNP-package}
\alias{ICSNP}
\docType{package}
\title{
Tools for Multivariate Nonparametrics
}
\description{
This package contains functions and tools for multivariate nonparametric tests and estimation.
}
\details{
\tabular{ll}{
Package: \tab ICSNP\cr
Type: \tab Package\cr
Version: \tab 1.0-5\cr
Date: \tab 2009-09-11\cr
License: \tab GPL (>= 2)\cr
}
This package contains tools for nonparametric multivariate analysis, including the estimation of location and shape as well as some tests for location and independece.
Shape matrices from this package can be used as one of the scatter matrices needed in the package \code{\link[ICS:ICS-package]{ICS}} whereas the tests of this package
can be used for testing in the framework of invariant coordinates or independent components obtained from the package \code{\link[ICS:ICS-package]{ICS}}.
The parametric Hotelling's T test serves as a reference for the nonparametric location tests.
}
\author{
Klaus Nordhausen, Seija \enc{Sirkiä}{Sirkia}, Hannu Oja and David E. Tyler

Maintainer:
Klaus Nordhausen, \email{klaus.nordhausen@uta.fi}
}


\keyword{package }
\keyword{multivariate}
\keyword{nonparametric}
\seealso{
\code{\link[ICS:ICS-package]{ICS}}
}
